% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirTest.R
\name{dirTest}
\alias{dirTest}
\title{Directed Test}
\usage{
dirTest(obj, null, conf.level = 0.95, ...)
}
\arguments{
\item{obj}{A fitted model object of class "flm", as returned by
\code{EstFLM}}

\item{null}{An optional global function to test against. The null
can be specified as a vector holding the discretized function object or,
alternatively, as an expansion of a given basis system. Defaults to
the zero function.}

\item{conf.level}{Confidence level (1 - significance level) used to decide
whether to reject or not.}

\item{...}{Further arguments passed to underlying C-routine. See source code of
\code{testseq}.}
}
\value{
An object of type \code{flm.test} with S3-methods \code{print} and
\code{summary}.
}
\description{
Test versus global alternative in Functional Linear Regression
}
\details{
\code{dirTest} performs the directed local test in the functional linear
regression model. This test can be used to discover a threshold t beyond which
the functional covariat have no significant effect on the outcome variable.
}
\references{
Rust, C. (2020) Directed Local Testing in Functional Linear Models. Available
from the author upon request.
}
\seealso{
EstFLM, GlobalTest
}
\author{
Christoph Rust
}
