% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstFLM.R
\name{EstFLM}
\alias{EstFLM}
\title{EstFLM}
\usage{
EstFLM(y, X, intercept = TRUE, type = "spline", df = NULL,
  rho = NULL, cpv = NULL)
}
\arguments{
\item{y}{Dependent variable, a numeric vector of length N.}

\item{X}{A N times p matrix holding the functional predictors
(one curve per row), assuming an equidistant grid of length
\code{p}.}

\item{intercept}{Logical, if TRUE, an intercept is included in the model.}

\item{model}{A list containing entries \code{type} (either
\code{"smoothspline"} or \code{"fpc"} and \code{df},
the latter controls flexibility of the fit (either number
of eigenfunctions or edf).}
}
\value{
A list containing estimation results with the
following entries:
\item{coefficients}{Estimated coefficients, a list
with entries \code{coefm} (coefficient matrix),
\code{beta} (beta function) and \code{beta_sd} (pointwise standard errors).
}
\item{residuals}{
 Residuals of the fit.
}
\item{fitted}{
 Fitted values
}
\item{model}{
 Parameters of the model fit.
}
}
\description{
Estimate a functional linear model
}
\author{
Christoph Rust
}
\keyword{FDA,}
\keyword{regression}
\keyword{smoothing}
\keyword{splines,}
