% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalTest.R
\name{GlobalTest}
\alias{GlobalTest}
\title{GlobalTest}
\usage{
GlobalTest(obj, null, type = "F", intercept = TRUE)
}
\arguments{
\item{obj}{A fitted model object of class "flm", as returned by
\code{EstFLM}}

\item{null}{An optional global function to test against. The null
can be specified as a vector holding the discretized function object or,
alternatively, as an expansion of a given basis system. Defaults to
the zero function.}

\item{type}{Which test to use. Currently, the following alternatives
are available: "F" (see Kong et al. 2017)}

\item{intercept}{TRUE or FALSE, indicating whether an intercept is
to be included in the alternative.}
}
\value{
An object of type \code{flm.test} with S3-methods \code{print} and
\code{summary}.
}
\description{
Test versus global alternative in Functional Linear Regression
}
\details{
\code{GlobalTest} performs the test of no association in the functional
linear model. The procedure can also be used to test whether the estimate
differs from a specific functional slope parameter.
}
\references{
Kong, D., Staicu, A., Maity, A. (2016) Classical Testing in Functional
Linear Models. Journal of Nonparametric Statistics, \emph{28}(4), 813-838.
}
\seealso{
EstFLM, GlobalTest
}
\author{
Christoph Rust
}
