% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natSplBasis.R
\name{natSplBasis}
\alias{natSplBasis}
\title{natSplBasis}
\usage{
natSplBasis(grd)
}
\arguments{
\item{grd}{A vector of length \eqn{p} indicating the grd values where the
spline functions are evaluated.}
}
\value{
A list with entries
\item{A_m}{
  The \eqn{p \times p} matrix \eqn{A} as described in (Crambes, Kneip and Sarda,
  2016)
}
\item{X_B}{
  A \eqn{p \times p} matrix of natural cubic spline basis evaluated at \code{grd}.
}
}
\description{
Natural Spline Basis and the Regularization Matrix \code{A_m}
}
\details{
\code{natSplBasis} calculates evaluated natural cubic splines
and the matrix \code{A_m} as described in Crambes, Kneip and Sarda
(2016).


\code{natSplBasis} calculates an evaluated natural cubic
spline basis corresponding to \code{grd} and a \eqn{p \times p} matrix
used for regularization in the smoothing splines estimator for the
functional linear model (Crambes, Kneip and Sarda,2016). \code{A_m} is
composed of a non-classical projection matrix and the classical
regularization matrix of squared second derivatives.
}
\examples{
\dontrun{

## Define Parameters for Simulation
p           <- 300
domain      <- c(0,1)
grd         <- seq(domain[1],domain[2],length.out = p)


NaturalSplines <- natSplBasis(grd)


}

}
\references{
Crambes, C., Kneip, A., Sarda, P. (2009) Smoothing Splines
Estimators for Functional Linear Regression. The Annals of
Statistics, \emph{37}(1), 35-72.
}
\author{
Christoph Rust
}
\keyword{models}
