% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqTest.R
\name{SeqTest}
\alias{SeqTest}
\title{SeqTest}
\usage{
SeqTest(obj, startval = 0, direction = "right", null)
}
\arguments{
\item{obj}{A fitted model object of class "flm", as returned by
\code{EstFLM}}

\item{startval}{A value in the domain \eqn{[0,1]} where it is
assumed that the effect is strongest. Defaults to 0.}

\item{direction}{Either of \code{"right","left","both"}, specifying
into which direction to test starting from \code{startvalue}.
Defaults to \code{"right"}.}

\item{null}{An optional global function to test against. The null
can be specified as a vector holding the discretized function object or,
alternatively, as an expansion of a given basis system. Defaults to
the zero function.}
}
\value{
An object of type \code{flm.dirtest} with S3-methods \code{print} and
\code{summary}.
}
\description{
Sequential test in the functional linear model
}
\details{
\code{SeqTest} performs a sequential test of no association in the
functional linear model in order to detect regions where the functions
have a significant influence on the scalar outcome. The procedure can
also be used to test whether the estimate differs from a specific
functional slope parameter.
}
\references{
Rust, C. (2020) Directed Local Testing in the Functional Linear Model.
Author Manuscript.
}
\seealso{
EstFLM, GlobalTest
}
\author{
Christoph Rust
}
