% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testseq.R
\name{testseq}
\alias{testseq}
\title{Test Sequence for directed Test}
\usage{
testseq(obj, null, startval, direction, gridvals = NULL, maxit = 1000L,
  tol = .Machine$double.eps^0.25)
}
\arguments{
\item{obj}{A fitted model object of class "flm", as returned by
\code{EstFLM}}

\item{null}{An optional global function to test against. The null
can be specified as a vector holding the discretized function object or,
alternatively, as an expansion of a given basis system. Defaults to
the zero function.}

\item{startval}{Grad value where the sequential test starts. Currently this
has no effect and the procedure starts at the left border of the domain.}

\item{direction}{One of "right", "left", "both". Specifies into which direction
from \code{startval} to procedure is applied.}

\item{gridvals}{Currently without effect.}

\item{maxit}{The maximum number of iterations in the root finding routine used
to relate the smoothing parameter \equation{\rho} and \code{df}.}

\item{tol}{Tolerance for the numerical routine used to relate the smoothing
parameter \equation{\rho} and \code{df}.}
}
\value{
A matrix with \code{p} rows and 7 columns. The first six columns
contain all the information used to calculate the test statistic and the
seventh contains the corresponding (approximate) p value.
}
\description{
This function is used by \code{dirTest} and calls the underlying C routine
\code{tstatseq}.
}
